requirements.txt
Django==4.0.4
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-extensions==3.2.3
drf-spectacular==0.22.1
Pillow==9.1.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pep8-naming==0.13.2
psycopg2-binary==2.9.10
django-filter==23.5
sqlparse==0.5.3
tzdata==2025.1
dotenv==0.9.9
